+++ packages/vsdx/lib/src/stencils.dart
@@ -1,5 +1,6 @@
 import 'dart:convert';
 import 'dart:math' as math;
+import 'dart:typed_data';
 
 import 'package:archive/archive.dart';
 import 'package:vsdx/vsdx.dart';